fixup wrapper with core refactor
